refactor: separate inventory and view packages by domain
