+++ modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/inventory/coroutine/VirtualCoroutineScope.kt
@@ -1,4 +1,4 @@
-package kr.hqservice.framework.inventory.coroutine.scope
+package kr.hqservice.framework.inventory.coroutine
 
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineName
