--- modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/inventory/element/ViewElement.kt
@@ -1,7 +1,0 @@
-package kr.hqservice.framework.inventory.element
-
-import kr.hqservice.framework.inventory.state.State
-
-sealed interface ViewElement {
-    fun subscribe(vararg states: State<*>)
-}
