+++ modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/view/HQViewModel.kt
@@ -1,12 +1,12 @@
-package kr.hqservice.framework.inventory
+package kr.hqservice.framework.view
 
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kr.hqservice.framework.coroutine.extension.BukkitMain
-import kr.hqservice.framework.inventory.coroutine.LifecycleOwner
-import kr.hqservice.framework.inventory.state.State
-import kr.hqservice.framework.inventory.state.impl.MutableStateFlowState
+import kr.hqservice.framework.view.coroutine.LifecycleOwner
+import kr.hqservice.framework.view.state.State
+import kr.hqservice.framework.view.state.impl.MutableStateFlowState
 import kotlin.coroutines.CoroutineContext
 
 abstract class HQViewModel : LifecycleOwner {
