+++ modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/view/element/ViewElement.kt
@@ -1,0 +1,7 @@
+package kr.hqservice.framework.view.element
+
+import kr.hqservice.framework.view.state.State
+
+sealed interface ViewElement {
+    fun subscribe(vararg states: State<*>)
+}
