+++ modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/view/navigator/Navigator.kt
@@ -1,0 +1,4 @@
+package kr.hqservice.framework.view.navigator
+
+interface Navigator {
+}
