+++ modules/bukkit-inventory/src/main/kotlin/kr/hqservice/framework/view/state/SubscribableState.kt
@@ -1,4 +1,4 @@
-package kr.hqservice.framework.inventory.state
+package kr.hqservice.framework.view.state
 
 import kotlinx.coroutines.flow.StateFlow
